pyproject.toml
[project]
name = "university-support-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.5",
    "pinecone>=7.3.0",
    "streamlit>=1.52.1",
    "tavily-python>=0.7.15",
]
